+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
     "pydantic>=2.10.5,<3.0.0",
     "sentence-transformers>=3.2.0,<4.0.0",
     "transformers>=4.44.2,<5.0.0",
-    "unstructured[md]>=0.16.0,<1.0.0",
+    "unstructured[md,pdf,docx]>=0.16.0,<1.0.0",
 ]
 
 [build-system]
