+++ src/synthetic_dataset_generator/apps/base.py
@@ -6,7 +6,7 @@ import argilla as rg
 import gradio as gr
 from datasets import Dataset, concatenate_datasets, load_dataset
 from gradio import OAuthToken
-from huggingface_hub import HfApi, upload_file
+from huggingface_hub import HfApi, upload_file, repo_exists
 
 from synthetic_dataset_generator.constants import MAX_NUM_ROWS
 from synthetic_dataset_generator.utils import get_argilla_client
@@ -159,3 +159,23 @@ def test_max_num_rows(num_rows: int) -> int:
             f"Number of rows is larger than the configured maximum. Setting number of rows to {MAX_NUM_ROWS}. Set environment variable `MAX_NUM_ROWS` to change this behavior."
         )
     return num_rows
+
+
+def get_iframe(hub_repo_id: str) -> str:
+    if not hub_repo_id:
+        return ""
+
+    if not repo_exists(repo_id=hub_repo_id, repo_type="dataset"):
+        return ""
+
+    url = f"https://huggingface.co/datasets/{hub_repo_id}/embed/viewer"
+    iframe = f"""
+    <iframe
+        src="{url}"
+        frameborder="0"
+        width="100%"
+        height="600px"
+    ></iframe>
+    """
+    return iframe
+
